requirements.txt
altgraph==0.17
async-timeout==3.0.1
attrs==21.2.0
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.4
et-xmlfile==1.1.0
idna==3.2
multidict==5.2.0
numpy==1.21.1
openpyxl==3.0.7
pandas==1.3.1
Pillow==8.3.1
pyinstaller==4.5.1
pyinstaller-hooks-contrib==2021.2
PySide2==5.15.2
PySimpleGUIQt==0.35.0
python-dateutil==2.8.2
pytz==2021.1
requests==2.26.0
shiboken2==5.15.2
six==1.16.0
typing-extensions==3.10.0.2
urllib3==1.26.6
yarl==1.7.0
